Revert "refactor(desktop): extract BuddyConnectionForm as shared component"
This reverts commit c6d4df19e5f151b7c9f7697c95b7bd4ac4d10640.
